+++ prog/dftb+/lib_dftbplus/main.F90
@@ -68,6 +68,7 @@ module dftbp_main
   use dftbp_mainio
   use dftbp_commontypes
   use dftbp_dispersions, only : TDispersionIface
+  use dftbp_solvation, only : TSolvation
   use dftbp_xmlf90
   use dftbp_thirdorder, only : TThirdOrder
   use dftbp_rangeseparated, only : TRangeSepFunc
@@ -449,13 +450,13 @@ contains
     end if
 
     if (tLatticeChanged) then
-      call handleLatticeChange(latVec, sccCalc, tStress, extPressure, cutOff%mCutOff, dispersion,&
+      call handleLatticeChange(latVec, sccCalc, tStress, extPressure, cutOff%mCutOff, dispersion, solvation, &
           & recVec, invLatVec, cellVol, recCellVol, extLatDerivs, cellVec, rCellVec)
     end if
 
     if (tCoordsChanged) then
       call handleCoordinateChange(env, coord0, latVec, invLatVec, species0, cutOff, orb,&
-          & tPeriodic, sccCalc, dispersion, thirdOrd, rangeSep, img2CentCell, iCellVec,&
+          & tPeriodic, sccCalc, dispersion, solvation, thirdOrd, rangeSep, img2CentCell, iCellVec,&
           & neighbourList, nAllAtom, coord0Fold, coord, species, rCellVec, nNeighbourSk,&
           & nNeighbourRep, nNeighbourLC, ham, over, H0, rhoPrim, iRhoPrim, iHam, ERhoPrim,&
           & iSparseStart, tPoisson)
@@ -495,7 +496,7 @@ contains
     call env%globalTimer%stopTimer(globalTimers%sparseH0S)
 
     if (tSetFillingTemp) then
-      call getTemperature(temperatureProfile, tempElec)
+      call temperatureProfile%getTemperature(tempElec)
     end if
 
     call electronicSolver%updateElectronicTemp(tempElec)
@@ -571,7 +572,7 @@ contains
       #:endif
 
         call addChargePotentials(env, sccCalc, qInput, q0, chargePerShell, orb, species,&
-            & neighbourList, img2CentCell, spinW, thirdOrd, potential, electrostatics, tPoisson,&
+            & neighbourList, img2CentCell, spinW, solvation, thirdOrd, potential, electrostatics, tPoisson,&
             & tUpload, shiftPerLUp)
 
         call addBlockChargePotentials(qBlockIn, qiBlockIn, tDftbU, tImHam, species, orb,&
@@ -667,7 +668,7 @@ contains
         call getChargePerShell(qOutput, orb, species, chargePerShell)
 
         call addChargePotentials(env, sccCalc, qOutput, q0, chargePerShell, orb, species,&
-            & neighbourList, img2CentCell, spinW, thirdOrd, potential, electrostatics,&
+            & neighbourList, img2CentCell, spinW, solvation, thirdOrd, potential, electrostatics,&
             & tPoissonTwice, tUpload, shiftPerLUp)
 
         call addBlockChargePotentials(qBlockOut, qiBlockOut, tDftbU, tImHam, species, orb,&
@@ -689,7 +690,7 @@ contains
 
       call getEnergies(sccCalc, qOutput, q0, chargePerShell, species, tExtField, isXlbomd,&
           & tDftbU, tDualSpinOrbit, rhoPrim, H0, orb, neighbourList, nNeighbourSk, img2CentCell,&
-          & iSparseStart, cellVol, extPressure, TS, potential, energy, thirdOrd, rangeSep,&
+          & iSparseStart, cellVol, extPressure, TS, potential, energy, thirdOrd, solvation, rangeSep,&
           & qDepExtPot, qBlockOut, qiBlockOut, nDftbUFunc, UJ, nUJ, iUJ, niUJ, xi,&
           & iAtInCentralRegion, tFixEf, Ef, onSiteElements)
 
@@ -738,7 +739,7 @@ contains
             & extPressure, cellVol, tAtomicEnergy, tDispersion, tEField, tPeriodic, nSpin, tSpin,&
             & tSpinOrbit, tSccCalc, allocated(onSiteElements), tNegf, invLatVec, kPoint,&
             & iAtInCentralRegion, electronicSolver, tDefinedFreeE, allocated(halogenXCorrection),&
-            & tRangeSep, allocated(thirdOrd))
+            & tRangeSep, allocated(thirdOrd), allocated(solvation))
       end if
 
       if (tConverged .or. tStopScc) then
@@ -764,6 +765,9 @@ contains
     if (isLinResp) then
       if (withMpi) then
         call error("Linear response calc. does not work with MPI yet")
+      end if
+      if (allocated(solvation)) then
+        call error("Solvation model do not work with linear response yet.")
       end if
       call ensureLinRespConditions(t3rd, tRealHS, tPeriodic, tCasidaForces)
       call calculateLinRespExcitations(env, lresp, parallelKS, sccCalc, qOutput, q0, over,&
@@ -825,7 +829,7 @@ contains
       call getGradients(env, sccCalc, tExtField, isXlbomd, nonSccDeriv, EField, rhoPrim, ERhoPrim,&
           & qOutput, q0, skHamCont, skOverCont, pRepCont, neighbourList, nNeighbourSk,&
           & nNeighbourRep, species, img2CentCell, iSparseStart, orb, potential, coord, derivs,&
-          & iRhoPrim, thirdOrd, qDepExtPot, chrgForces, dispersion, rangeSep, SSqrReal, over,&
+          & iRhoPrim, thirdOrd, solvation, qDepExtPot, chrgForces, dispersion, rangeSep, SSqrReal, over,&
           & denseDesc, deltaRhoOutSqr, tPoisson, halogenXCorrection)
 
       if (tCasidaForces) then
@@ -842,7 +846,7 @@ contains
             & q0, skHamCont, skOverCont, pRepCont, neighbourList, nNeighbourSk, nNeighbourRep,&
             & species, img2CentCell, iSparseStart, orb, potential, coord, latVec,&
             & invLatVec, cellVol, coord0, totalStress, totalLatDeriv, intPressure, iRhoPrim,&
-            & dispersion, halogenXCorrection)
+            & solvation, dispersion, halogenXCorrection)
         call env%globalTimer%stopTimer(globalTimers%stressCalc)
         call printVolume(cellVol)
 
@@ -1094,7 +1098,7 @@ contains
 
 
   !> Does the operations that are necessary after a lattice vector update
-  subroutine handleLatticeChange(latVecs, sccCalc, tStress, extPressure, mCutOff, dispersion,&
+  subroutine handleLatticeChange(latVecs, sccCalc, tStress, extPressure, mCutOff, dispersion, solvation, &
       & recVecs, recVecs2p, cellVol, recCellVol, extLatDerivs, cellVecs, rCellVecs)
 
     !> lattice vectors
@@ -1114,6 +1118,9 @@ contains
 
     !> Dispersion interactions object
     class(TDispersionIface), allocatable, intent(inout) :: dispersion
+
+    !> Solvation model
+    class(TSolvation), allocatable, intent(inout) :: solvation
 
     !> Reciprocal lattice vectors
     real(dp), intent(out) :: recVecs(:,:)
@@ -1154,6 +1161,10 @@ contains
       call dispersion%updateLatVecs(latVecs)
       mCutOff = max(mCutOff, dispersion%getRCutOff())
     end if
+    if (allocated(solvation)) then
+      call solvation%updateLatVecs(latVecs)
+      mCutOff = max(mCutOff, solvation%getRCutOff())
+    end if
     call getCellTranslations(cellVecs, rCellVecs, latVecs, recVecs2p, mCutOff)
 
   end subroutine handleLatticeChange
@@ -1161,7 +1172,7 @@ contains
 
   !> Does the operations that are necessary after atomic coordinates change
   subroutine handleCoordinateChange(env, coord0, latVec, invLatVec, species0, cutOff, orb,&
-      & tPeriodic, sccCalc, dispersion, thirdOrd, rangeSep, img2CentCell, iCellVec, neighbourList,&
+      & tPeriodic, sccCalc, dispersion, solvation, thirdOrd, rangeSep, img2CentCell, iCellVec, neighbourList,&
       & nAllAtom, coord0Fold, coord, species, rCellVec, nNeighbourSK, nNeighbourRep, nNeighbourLC,&
       & ham, over, H0, rhoPrim, iRhoPrim, iHam, ERhoPrim, iSparseStart, tPoisson)
 
@@ -1196,6 +1207,9 @@ contains
 
     !> Dispersion interactions
     class(TDispersionIface), allocatable, intent(inout) :: dispersion
+
+    !> Solvation model
+    class(TSolvation), allocatable, intent(inout) :: solvation
 
     !> Third order SCC interactions
     type(TThirdOrder), allocatable, intent(inout) :: thirdOrd
@@ -1306,6 +1320,9 @@ contains
 
     if (allocated(dispersion)) then
       call dispersion%updateCoords(env, neighbourList, img2CentCell, coord, species0)
+    end if
+    if (allocated(solvation)) then
+      call solvation%updateCoords(env, neighbourList, img2CentCell, coord, species0)
     end if
     if (allocated(thirdOrd)) then
       call thirdOrd%updateCoords(neighbourList, species)
@@ -1791,7 +1808,7 @@ contains
 
   !> Add potentials comming from point charges.
   subroutine addChargePotentials(env, sccCalc, qInput, q0, chargePerShell, orb, species,&
-      & neighbourList, img2CentCell, spinW, thirdOrd, potential, electrostatics, tPoisson,&
+      & neighbourList, img2CentCell, spinW, solvation, thirdOrd, potential, electrostatics, tPoisson,&
       & tUpload, shiftPerLUp)
 
     !> Environment settings
@@ -1823,6 +1840,9 @@ contains
 
     !> spin constants
     real(dp), intent(in), allocatable :: spinW(:,:,:)
+
+    !> Solvation mode
+    class(TSolvation), allocatable, intent(inout) :: solvation
 
     !> third order SCC interactions
     type(TThirdOrder), allocatable, intent(inout) :: thirdOrd
@@ -1907,6 +1927,13 @@ contains
     if (allocated(thirdOrd)) then
       call thirdOrd%updateCharges(pSpecies0, neighbourList, qInput, q0, img2CentCell, orb)
       call thirdOrd%getShifts(atomPot(:,1), shellPot(:,:,1))
+      potential%intAtom(:,1) = potential%intAtom(:,1) + atomPot(:,1)
+      potential%intShell(:,:,1) = potential%intShell(:,:,1) + shellPot(:,:,1)
+    end if
+
+    if (allocated(solvation)) then
+      call solvation%updateCharges(env, pSpecies0, neighbourList, qInput, q0, img2CentCell, orb)
+      call solvation%getShifts(atomPot(:,1), shellPot(:,:,1))
       potential%intAtom(:,1) = potential%intAtom(:,1) + atomPot(:,1)
       potential%intShell(:,:,1) = potential%intShell(:,:,1) + shellPot(:,:,1)
     end if
@@ -3391,7 +3418,7 @@ contains
   !> Calculates various energy contribution that can potentially update for the same geometry
   subroutine getEnergies(sccCalc, qOrb, q0, chargePerShell, species, tExtField, isXlbomd, tDftbU,&
       & tDualSpinOrbit, rhoPrim, H0, orb, neighbourList, nNeighbourSK, img2CentCell, iSparseStart,&
-      & cellVol, extPressure, TS, potential, energy, thirdOrd, rangeSep, qDepExtPot, qBlock,&
+      & cellVol, extPressure, TS, potential, energy, thirdOrd, solvation, rangeSep, qDepExtPot, qBlock,&
       & qiBlock, nDftbUFunc, UJ, nUJ, iUJ, niUJ, xi, iAtInCentralRegion, tFixEf, Ef, onSiteElements)
 
     !> SCC module internal variables
@@ -3459,6 +3486,9 @@ contains
 
     !> 3rd order settings
     type(TThirdOrder), intent(inout), allocatable :: thirdOrd
+
+    !> Solvation model
+    class(TSolvation), allocatable, intent(inout) :: solvation
 
     !> Data from rangeseparated calculations
     type(TRangeSepFunc), intent(inout), allocatable ::rangeSep
@@ -3548,6 +3578,11 @@ contains
       energy%e3rd = sum(energy%atom3rd(iAtInCentralRegion(:)))
     end if
 
+    if (allocated(solvation)) then
+      call solvation%getEnergies(energy%atomSolv)
+      energy%eSolv = sum(energy%atomSolv(iAtInCentralRegion(:)))
+    end if
+
     if (allocated(onSiteElements)) then
       call getEons(energy%atomOnSite, qBlock, qiBlock, q0, onSiteElements, species, orb)
       energy%eOnSite = sum(energy%atomOnSite)
@@ -3570,7 +3605,7 @@ contains
     end if
 
     energy%Eelec = energy%EnonSCC + energy%ESCC + energy%Espin + energy%ELS + energy%Edftbu&
-        & + energy%Eext + energy%e3rd + energy%eOnSite
+        & + energy%Eext + energy%e3rd + energy%eOnSite + energy%ESolv
 
     !> Add exchange conribution for range separated calculations
     if (allocated(rangeSep)) then
@@ -3580,7 +3615,8 @@ contains
     end if
 
     energy%atomElec(:) = energy%atomNonSCC + energy%atomSCC + energy%atomSpin + energy%atomDftbu&
-        & + energy%atomLS + energy%atomExt + energy%atom3rd + energy%atomOnSite
+        & + energy%atomLS + energy%atomExt + energy%atom3rd + energy%atomOnSite &
+        & + energy%atomSolv
     energy%atomTotal(:) = energy%atomElec + energy%atomRep + energy%atomDisp + energy%atomHalogenX
     energy%Etotal = energy%Eelec + energy%Erep + energy%eDisp + energy%eHalogenX
     energy%EMermin = energy%Etotal - sum(TS)
@@ -5182,7 +5218,7 @@ contains
   !> Calculates the gradients
   subroutine getGradients(env, sccCalc, tExtField, isXlbomd, nonSccDeriv, EField, rhoPrim, ERhoPrim,&
       & qOutput, q0, skHamCont, skOverCont, pRepCont, neighbourList, nNeighbourSK, nNeighbourRep,&
-      & species, img2CentCell, iSparseStart, orb, potential, coord, derivs, iRhoPrim, thirdOrd,&
+      & species, img2CentCell, iSparseStart, orb, potential, coord, derivs, iRhoPrim, thirdOrd, solvation,&
       & qDepExtPot, chrgForces, dispersion, rangeSep, SSqrReal, over, denseDesc, deltaRhoOutSqr,&
       & tPoisson, halogenXCorrection)
 
@@ -5260,6 +5296,9 @@ contains
 
     !> Is 3rd order SCC being used
     type(TThirdOrder), intent(inout), allocatable :: thirdOrd
+
+    !> Solvation model
+    class(TSolvation), allocatable, intent(inout) :: solvation
 
     !> Population dependant external potential
     type(TQDepExtPotProxy), intent(inout), allocatable :: qDepExtPot
@@ -5384,6 +5423,14 @@ contains
     
     end if !tPoisson
 
+    if (allocated(solvation)) then
+      if (isXlbomd) then
+        call error("XLBOMD does not work with solvation yet!")
+      else
+        call solvation%addGradients(env, neighbourList, species, coord, img2CentCell, derivs)
+      end if
+    end if
+
     if (allocated(dispersion)) then
       call dispersion%addGradients(derivs)
     end if
@@ -5467,7 +5514,7 @@ contains
   subroutine getStress(env, sccCalc, thirdOrd, tExtField, nonSccDeriv, rhoPrim, ERhoPrim, qOutput,&
       & q0, skHamCont, skOverCont, pRepCont, neighbourList, nNeighbourSk, nNeighbourRep, species,&
       & img2CentCell, iSparseStart, orb, potential, coord, latVec, invLatVec, cellVol, coord0,&
-      & totalStress, totalLatDeriv, intPressure, iRhoPrim, dispersion, halogenXCorrection)
+      & totalStress, totalLatDeriv, intPressure, iRhoPrim, solvation, dispersion, halogenXCorrection)
 
     !> Environment settings
     type(TEnvironment), intent(in) :: env
@@ -5556,6 +5603,9 @@ contains
     !> imaginary part of the density matrix (if present)
     real(dp), intent(in), allocatable :: iRhoPrim(:,:)
 
+    !> Solvation model
+    class(TSolvation), allocatable, intent(inout) :: solvation
+
     !> dispersion interactions
     class(TDispersionIface), allocatable, intent(inout) :: dispersion
 
@@ -5591,6 +5641,11 @@ contains
             & skOverCont, coord, species, neighbourList%iNeighbour, nNeighbourSK, img2CentCell,&
             & iSparseStart, orb, cellVol)
       end if
+    end if
+
+    if (allocated(solvation)) then
+      call solvation%getStress(tmpStress)
+      totalStress(:,:) = totalStress + tmpStress
     end if
 
     if (allocated(dispersion)) then
@@ -5994,7 +6049,7 @@ contains
     velocities(:, indMovedAtom) = movedVelo(:,:)
 
     if (allocated(temperatureProfile)) then
-      call next(temperatureProfile)
+      call temperatureProfile%next()
     end if
     call evalKE(energy%Ekin, movedVelo, movedMass(1,:))
     call evalkT(pMdFrame, tempIon, movedVelo, movedMass(1,:))