+++ CHANGELOG.rst
@@ -44,6 +44,7 @@ Fixed
 - Correct update of block mulliken population for onsite correction with
   range-separation hybrid DFTB.
 
+- MD temperature profiles that do not start with an initial constant temperature
 
 1.02 (2019-01-18)
 =================